+++ +sbfsem/+builtin/GCLBoundary.m
@@ -5,6 +5,8 @@ classdef GCLBoundary < sbfsem.core.BoundaryMarker
 		function obj = GCLBoundary(source)
 			obj@sbfsem.core.BoundaryMarker(source);
 			obj.TYPEID = 235;
+            
+            obj.update();
 		end
 	end
 end